# Remove useless template instantiations